Generalize "heading" syntax matching.
Some sites extend their arcanist diff templates to include additional
headings (e.g. "JIRA Issues:"). These could be added to local syntax
extensions, but it seems more natural to generalize this plugin's notion
of a "heading".

This is (intentionally) less specific than the previous approach and
carries with it the following shortcomings:

 1. Arbitrary commit message text (e.g. /^Foo:/ could be matched).
 2. Individual headings can no longer have their own highlighting.

... but those feel like worthwhile tradeoffs.